requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
django-jazzmin>=3.0
daphne>=4.0
whitenoise>=6.0
Pillow>=10.0


